platform/Linux: fix GCC 11 ARM build (nested designated-init + uint64_t format)
GCC 11.2.0 (kirkstone ARMv7 toolchain) rejects nested designated initializers
into anonymous unions — a bug fixed in GCC 12. Replace all occurrences with
sequential field assignment (same semantics, compiles on GCC 11+).

Files: BLEManagerImpl, PlatformManagerImpl, ThreadStackManagerImpl,
ConnectivityManagerImpl_{NetworkManagement,WiFiPaf}WpaSupplicant,
GenericConnectivityManagerImpl_Thread.ipp, DeviceControlServer,
FailSafeContext, BindingCluster.

Also fix -Werror=format= in pushav-clip-recorder and pushav-transport:
uint64_t is 'long long' on ARM32, not 'long'; cast to unsigned long long
with %llu.

+++ examples/camera-app/linux/src/pushav-clip-recorder.cpp
@@ -106,8 +106,8 @@ PushAVClipRecorder::~PushAVClipRecorder()
     if (IsFileReadyForUpload(mpdPath))
     {
         UpdateMPDParams(mpdPath);
-        ChipLogProgress(Camera, "Uploading final MPD: %s for track: %s, sessionID: %lu, connectionID: %u", mpdPath.c_str(),
-                        mClipInfo.mTrackName.c_str(), mClipInfo.mSessionNumber, mConnectionID);
+        ChipLogProgress(Camera, "Uploading final MPD: %s for track: %s, sessionID: %llu, connectionID: %u", mpdPath.c_str(),
+                        mClipInfo.mTrackName.c_str(), static_cast<unsigned long long>(mClipInfo.mSessionNumber), mConnectionID);
         CheckAndUploadFile(mpdPath.string());
     }
 }
@@ -266,7 +266,7 @@ AVPacket * PushAVClipRecorder::CreatePacket(const uint8_t * data, int size, int6
                 mClipInfo.mClipStartPTS = timestampMs;
             packet->flags = AV_PKT_FLAG_KEY;
 
-            ChipLogProgress(Camera, "Found I-frame at timestamp: %ld ms", timestampMs);
+            ChipLogProgress(Camera, "Found I-frame at timestamp: %lld ms", static_cast<long long>(timestampMs));
         }
 
         if (mClipInfo.mHasVideo && mClipInfo.mClipStartPTS == 0)
